feat: added context and moved main api all there
